+++ test/sql/cleanup.sql
@@ -1,6 +1,8 @@
 -- cleanup.sql: Clean up the regression test environment, in particular roles 
 --              (all components inside the regression database will be deleted with the regression database)
 
+\! rm -Rf /tmp/emaj_test
+
 -----------------------------
 -- drop emaj_regression_tests_adm_user role
 -----------------------------